refactor(functions): separate firebase users utils
