+++ packages/functions/src/api/updateTwUsers.ts
@@ -2,7 +2,7 @@ import { WatchData } from '@yukukuru/types';
 import * as Twitter from 'twitter';
 import { firestore } from '../modules/firebase';
 import { env } from '../utils/env';
-import { updateUserLastUpdatedTwUsers } from '../utils/firestore';
+import { updateUserLastUpdatedTwUsers } from '../utils/firestore/users/updateUserLastUpdatedTwUsers';
 import { setTwUsers } from '../utils/firestore/twUsers/setTwUsers';
 import { getUsersLookup } from '../utils/twitter';
